test(ts): unify `Add User` benchmark
